pam: merge pam-plugin-* packages into libpam-modules
pam-plugin-* are automatically generated and include
corresponding *.so, but in Debian, they don't exist and
provided by libpam-modules instead in Debian.

The advantages of splitting libpam-modules into pam-plugin-*
may be less than providing only libpam-modules package.
For examples, there are not so much ${base_libdir}/security/*.so
and they are not so big. Also, selecting required pam-plugins
in recipes that depend on some plugins is little troublesome.

Therefore, we merge plugin packages into libpam-modules again to
simplify package implementation.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>