busybox: support installing inittab and rcS
If busybox is the init manager, inittab and rcS are essential
for normal boot. By default, the minimal inittab and rcS are
installed by the busybox recipe. They can be replaced easily
by adding a local directory to FILESEXTRAPATHS_prepend and
putting inittab or rcS files into the directory.

Respawns to getty tty1 or ttys in SERIAL_CONSOLES are automatically
added to inittab. The ttys can be replaced by setting
BUSYBOX_INITTAB_GETTYS with the same format as SERIAL_CONSOLES.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>