busybox: support appending .cfgs to defconfig in do_configure
If .cfg files are included in SRC_URI, they are automatically
appended to defconfig. This function is implemented in
the original busybox recipe, so this commit just restores it.

This function is useful for changing a few configs in defconfig,
and has an advantage that the differences between defconfig and
the final .config are kept in separated .cfg files.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>