ntp: include all suitable non-PARSE and PARSE clocks and enable linuxcaps
Add options "--enable-all-clocks --enable-parse-clocks" as debian/rules.

Enable linuxcaps for fixing error when start NTP server:
  | The ``user'' option has been disabled -- built without --enable-clockctl or --enable-linuxcaps

Debian does not add "--enable-linuxcaps" directly to debian/rules,
but according to debian/control, ntp depends on libcap2-dev for building,
so we use "--enable-linuxcaps" instead of "--enable-clockctl".

"pps-tools" is not required for option "--enable-all-clocks", so remove it.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>