dhcp: Correct files shipment and LIC_FILES_CHKSUM
The append packages need be added with higher priority than default packages,
if not, files of append packages will be stolen by default packages.

The package names in "PACKAGES" and "FILES_" should be same,
both are "isc-dhcp-*" or both are "isc-${PN}-*",
not one is "isc-dhcp-*", but other is "isc-${PN}-*".

LIC_FILES_CHKSUM should get checksum of the whole file LICENSE, not only from line 4,
because line 1 and line 2 also contain copyright information.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>