SDK: install native ncurses header and library for cross menuconfig
In make menuconfig of kernel and busybox, which normaly assumes
native build, scripts/kconfig/lxdialog/check-lxdialog.sh checks
ncurses headers and libraries in the host system with pkg-config,
native gcc, or by finding header and library files installed
in the host system.

If we cross-build kernel or busybox with SDK,
check-lxdialog.sh -ccflags and -ldflags returns invalid flags
because pkg-config called in check-lxdialog.sh returns
state of ncurses in the target sysroot in SDK, and others
return state of ncurses in the host system.

Yocto Project avoids this problem by applying the following patches
to their kernel and busybox source code.
http://git.yoctoproject.org/cgit/cgit.cgi/linux-yocto-4.4/commit?id=badf6fedf455958fe0ff3c060c8e3965ef6d80dc
http://git.yoctoproject.org/cgit/cgit.cgi/linux-yocto-4.4/commit?id=26e282c0686e56cfc1c7eb4cc4c63275e49aad6e
http://git.yoctoproject.org/cgit/cgit.cgi/poky/commit/?id=da6a4f6c2b2f108e52fcb546d68c8e16ec546b6c

We also use the same approach;
1. Always install ncurses development package into native sysroot in SDK
2. Export the following four variables, which are also exported in
   cml1.bbclass, so that check-lxdialog.sh in cross-build with SDK properly
   finds native ncurses header and library from native sysroot in SDK.
   HOST_EXTRACFLAGS, HOSTLDFLAGS, CROSS_CURSES_LIB, CROSS_CURSES_INC

NOTE: The above three patches to handle CROSS_CURSES_LIB and
CROSS_CURSES_INC must be applied to kernel or busybox source code
that you are trying to cross-build with SDK. Without these patches,
make menuconfig (check-lxdialog.sh) still depends on the host system
or the target sysroot in SDK. This also means that currently there is
no way to fix the problem without patching kernel or busybox source code.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>