recipes: Set PV to source code version
Stop using PV as "git" and specify it to the true version of source code.

Add a new task do_debian_verify_version to check if PV and version
from source code changelog are match.
The source code version is parsed base on the way of cdbs
used to parse DEB_UPSTREAM_VERSION from debian/changelog.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>