libpam: Handle selinux option by DISTRO_FEATURES
In pam config file of openssh and shadow, there is a rule requires pam_selinux.so.
We need check whether libpam is built with option selinux or not
to disable this rule.

It's better to enable/disable selinux by using DISTRO_FEATURES,
which let openssh and shadow know if pam_selinux.so will be built.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>
Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>