kbd: remove doc/utf directory
This directory contains odd file named in UTF-8.
The file causes an error in do_debian_fix_timestamp(),
so I want to remove it beforehand.
bash regards the length of the filename as 2, however, dash does as 6.
It is so difficult to specify the name of the file name to remove,
I decided to remove the entire directory.

Signed-off-by: Toshiko YOSHIDA <toshiko.yoshida@toshiba.co.jp>
Signed-off-by: Masahiro Yamada <masahiro17.yamada@toshiba.co.jp>