grub: 'grub_build_mkfont_excuse' has a dummy string
If unifont-related packages are installed on a host,
such as xfonts-unifont, ttf-unifont and psf-unifont,
and if there are no unifont-related packages on a
target, building grub2 for the target fails.
This patch forces to tell configure file that there are
no unifont packages on the host even if they really exist.

Signed-off-by: Daichi Fukui <daichi1.fukui@toshiba.co.jp>