busybox: enable CLOCAL for getty command in inittab
The issue is that less command for ttyS* is blocked
if the CLOCAL is not set by getty.
busybox getty doen't set CLOCAL if -L flag is not given.
This seems to be busybox getty specific behavior (for example,
agetty of util-linux automatically inherits CLOCAL state of kernel).

There is no clear reason for disabling CLOCAL, so
-L should be set in the default inittab of our busybox.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>