cups: Keep CUPS_SERVERBIN as default path
The desired install path is /usr/lib/cups which is same as Debian.

We used a patch from 'meta' which made CUPS_SERVERBIN change path to "${libdir}/cups",
but 'libdir' was changed to /usr/lib/<multiarch-triplet>.

Remove this patch and keep using the default value of cups's config
to make installation path correct.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>