python/python3: Remove unused multilib patches
The FHS and LSB have standardized the x86_64 architecture to use "/lib64"
as the path for 64-bit x86 libraries, with "/lib" reserved for 32-bit x86 libraries.
Poky follows this standard and applies patches to python to satisfy it.

However, meta-debian follows Debian standard, so these patches have no use for us.
We want to keep the code base same as Debian as possible, so remove these patches.

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>