python: Do not move libpython for nativesdk class
When we use toolchain and run python, we got a issue look like:
    | $PATH_TO_SDK/sysroots/i686-skeliossdk-linux/usr/bin/python2.7: error while loading
    | shared libraries: libpython2.7.so.1.0: cannot open shared object file:
    | No such file or directory

The reason is that the library is moved to ${libdir}/${DEB_HOST_MULTIARCH}/ and
the binary file cannot find the library. We should just move the lib with class target
and keep them into ${libdir} with nativesdk class.

Signed-off-by: Hoang Van Tuyen <tuyen.hoangvan@toshiba-tsdv.com>
Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>