classes: Add debian-package-ng
This class supports retrieval of source code from snapshot.debian.org.
In many cases, recipe building will be done by specifying version containing
Debian version in package version (PV). And if the version contains epoch,
it is corresponded by specifying it as DPV variable.

NOTE:
This version does not automatically process the hash of each file.Recipe
mainatainer needs to list the hash of the file in the recipe.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu@miraclelinux.com>
Signed-off-by: Nobuhiro Iwamatsu <iwamatsu@nigauri.org>