debian-package.bbclass: Update checking DEBIAN_QUILT_PATCHES
In the Debian source package, debian/patches directory may be empty, or
debian/parches/series with no data.
The current process checks the contents of DEBIAN_QUILT_PATCHES and the
existence of a directory, but in the above case an error occurs.

This fix switched to the check process of debian/parches/series file.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu@miraclelinux.com>