classes: Change how to get debian source package name
Currently, debian-source.bbclass gets debian source package name and specifies
it in DSPN variable. This variable is only used for processing of snapshot.d.o.
I think it is not good to add variables carelessly.

This patch adds the function to get the debian source package name from the
file path in DEBIAN_SRC_URI.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro1.iwamatsu@toshiba.co.jp>