openssl: ptest: Do not change target file name
OpenSSL's ptest failed by following errors.

test/recipes/90-test_shlibload.t ..
1..10
Failed to load libcrypto
../util/shlib_wrap.sh ./shlibloadtest -crypto_first ../libcrypto.so.1.1 ../libssl.so.1.1 /tmp/w5Gepyd2VU => 1
not ok 1 - running shlibloadtest -crypto_first /tmp/w5Gepyd2VU
not ok 2
Failed to load libssl

Test test/recipes/90-test_shlibload.t runs following command.
../util/shlib_wrap.sh ./shlibloadtest -crypto_first ../libcrypto.so.1.1 ../libssl.so.1.1 /tmp/w5Gepyd2VU => 1

This test uses libcrypto.so.1.1 and libssl.so.1.1 for tests which are in /usr/lib/openssl/ptest/ but
we have libcrypto.so and libssl.so in the directory not libcrypto.so.1.1 and libssl.so.1.1.
It seems it is the root cause of test failure.
Therefore in the do_install_ptest(), create symbolic links with same name.

This patch creates symbolic links as poky does.