kernel: support qemu's vga,edid option
To use qemu with gui, VGA and edid option is required.
Without correct kernel configuration, qemu shows "Guest as not
initialized the display" on its display. So that we cannot
use qemu with gui.

However, remove "-device VGA,edid=on" from qemu option, it works
fine for console login, but it's not support xwindow system. That
means we cannot use qemu for testing applictions which run on
xwindow system.

This patch add kernel options to run qemu with "-device VGA,edid=on"
option.