dbus-test: Improve ptest dependencies
The dbus-test package is empty, move its dependencies to the ${PN}-ptest
package. Also ensure that it doesn't depend on the empty  package
which is about to start causing image failures in the following commit.
In this case the correct dependency is dbus itself.

(Base on OE-Core rev: a2915db76d6a2007b22b20f945498c84e7fa8132)

Signed-off-by: Trung Do <trung.dothanh@toshiba-tsdv.com>