Add recipes for making SDK.
We think that the SDK must have kernel sources.
If not, we can't compile any kernel modules / kernel drivers using
this SDK.
So we add recipes for the SDK.

Furthermore we execute "make scripts" after extracting kernel source.
Because there is no command needed for kernel module building.
We have to execute this command at least one time after extracting kernel source.
So we do this task automatically when installing SDK.