dash: Add dash package
This commit adds dash package which is contained in CIP's package
list[1].
So, it'd be nice to have dash package in meta-debian.

Tested on deby-tiny(qemuarm64), that worked as expected.

```
 Debian GNU/Linux 10 192.168.7.2 /dev/ttyAMA0

 192.168.7.2 login: root

 The programs included with the Debian GNU/Linux system are free software;
 the exact distribution terms for each program are described in the
 individual files in /usr/share/doc/*/copyright.

 Debian GNU/Linux comes with ABSOLUTELY NO WARRANTY, to the extent
 permitted by applicable law.
 login[1385]: root login on 'ttyAMA0'
 # echo $SHELL
 /bin/sh
 # ls -la /bin/sh
 lrwxrwxrwx    1 root     root             9 Oct 28 01:53 /bin/sh -> /bin/dash
```

Recipe is based on meta-openembedded's one[2].

The dash package sets ALTERNATIVE_PRIORITY to 110 which is greater than
bash. It was discussed what priority is good on github then decided to
have bigger number than bash[3].

1: https://gitlab.com/cip-project/cip-core/cip-pkglist/-/blob/master/pkglist_buster.yml
2:
http://cgit.openembedded.org/meta-openembedded/tree/meta-oe/recipes-shells/dash/dash_0.5.11.2.bb?h=master
3: https://github.com/meta-debian/meta-debian/pull/246

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>