debian-check-aarea.bbclass: Add do_debian_check_aarea()
do_debian_check_aarea() checks if the archive area
(main, contrib, non-free) of each source package and raises an error
if the archive area is either contrib or non-free and it's not
included in DEBIAN_ALLOWED_ARCHIVE_AREA (empty by default).

This is mainly intended to avoid using non-free packages
unintentionally.

Signed-off-by: Kazuhiro Hayashi <zuka0828@gmail.com>