squashfs-tools: Update license information
In 4.2+20130409-2, the source tree is restructured significantly and
the main license file (COPYING) and other files like
CHANGES, INSTALL, etc. in the top directoly are removed.

Updated recipe checks MD5 of the header text of squashfs_fs.h
instead of COPYING, which may be one of the core files
of this package.

Signed-off-by: Kazuhiro Hayashi <zuka0828@gmail.com>