Remove pkgconf
`pkgconf` and `pkg-config` conflict each other in Debian as well.

In meta-debian, `pkg-config` (`pkgconfig` reipce) is still mainly
used as a provider of this feature. It looks that there is no recipe
that depend on `pkgconf` (or binary packages built from that),
so it should be safe to be removed.

Signed-off-by: Kazuhiro Hayashi <zuka0828@gmail.com>