gcc: Add CVE-2021-37322 to CVE_CHECK_WHITELIST
CVE-2021-37322 is labeled as vulnerability of gcc on the NVD information.
https://nvd.nist.gov/vuln/detail/CVE-2021-37322
But this is labeled as vulnerability of binutils on Debian security tracker.
https://security-tracker.debian.org/tracker/CVE-2021-37322

This is because same module is used on both packages.
As for debian based system, real installed module is made from binutils.
And CVE-2021-37322 was fixed on binutils package and same fixed code could be found on gcc package.
So we add CVE-2021-37322 to CVE_CHECK_WHITELIST of gcc.