openssh: Add run-ptest file from OE-Core
This patch resolves the make command message below output when running ptest.
  make: Target 'tests' not remade because of errors.

That message output cause by the unittest failing of ptest.
(The unittest runs by 'unit' target of regress/Makefile)
Because, the unittest binaries does not exist.
The recipe's do_compile_ptest has the following comment:
    # skip regress/unittests/ binaries: this will silently skip
    # unittests in run-ptests which is good because they are so slow.

The purpose of this patch is to add the definition of the SKIP_UNIT variable to
run-ptest in order to skip unittest in ptest.
The SKIP_UNIT definition was not implemented in Poky, the run-ptest file was
backport from OE-Core rev: 9f031d936edb69c038639656288c84564434b906

This run-ptest file is equivalent to Poky's run-ptest file with the following
OE-Core patches.
  - 9f031d936edb69c038639656288c84564434b906
  - d3ffbebf43c23faa43af81c9ecf6fcaef36d675b